Enemys now shows damage taken.

+++ enemy.h
@@ -17,7 +17,7 @@ private:
 	int counterWalking;
 	int direction; // 1 - up, 2 - down, 3 - left, 4 - right
 	int counter;
-	float hp;
+	int hp;
 	bool alive;
 	int offset;
 	CircleShape circle;
